requirements.txt
google-cloud-bigquery==3.4.1
Flask == 2.2.2